Move scanner.Combiner/Mapper to package zbuf
The purpose of package scanner is to provide a proc interface
for a zbuf.Reader source. As such scanner.Combiner/Mapper to fit
within the package. Move them to the growing library of truly useful
generic zng stream operators in zbuf.
