+++ cmd/zq/zq.go
@@ -15,7 +15,6 @@ import (
 	"github.com/brimsec/zq/emitter"
 	"github.com/brimsec/zq/pkg/nano"
 	"github.com/brimsec/zq/proc"
-	"github.com/brimsec/zq/scanner"
 	"github.com/brimsec/zq/zbuf"
 	"github.com/brimsec/zq/zio"
 	"github.com/brimsec/zq/zio/detector"
@@ -230,7 +229,7 @@ func (c *Command) Run(args []string) error {
 			readers[i] = zbuf.NewWarningReader(r, wch)
 		}
 	}
-	reader := scanner.NewCombiner(readers)
+	reader := zbuf.NewCombiner(readers)
 	defer reader.Close()
 
 	writer, err := c.openOutput()
