+++ zbuf/combiner.go
@@ -1,39 +1,24 @@
-package scanner
+package zbuf
 
 import (
 	"fmt"
 	"io"
 
-	"github.com/brimsec/zq/zbuf"
-	"github.com/brimsec/zq/zio/detector"
 	"github.com/brimsec/zq/zng"
-	"github.com/brimsec/zq/zng/resolver"
 )
 
 type Combiner struct {
-	readers []zbuf.Reader
+	readers []Reader
 	hol     []*zng.Record
 	done    []bool
 }
 
-func NewCombiner(readers []zbuf.Reader) *Combiner {
+func NewCombiner(readers []Reader) *Combiner {
 	return &Combiner{
 		readers: readers,
 		hol:     make([]*zng.Record, len(readers)),
 		done:    make([]bool, len(readers)),
 	}
-}
-
-func OpenFiles(zctx *resolver.Context, paths ...string) (*Combiner, error) {
-	var readers []zbuf.Reader
-	for _, path := range paths {
-		reader, err := detector.OpenFile(zctx, path, detector.OpenConfig{})
-		if err != nil {
-			return nil, err
-		}
-		readers = append(readers, reader)
-	}
-	return NewCombiner(readers), nil
 }
 
 func (c *Combiner) Read() (*zng.Record, error) {
@@ -65,14 +50,14 @@ func (c *Combiner) Read() (*zng.Record, error) {
 	return tup, nil
 }
 
-func (c *Combiner) closeReader(r zbuf.Reader) error {
+func (c *Combiner) closeReader(r Reader) error {
 	if closer, ok := r.(io.Closer); ok {
 		return closer.Close()
 	}
 	return nil
 }
 
-// Close closes underlying zbuf.Readers implementing the io.Closer
+// Close closes underlying Readers implementing the io.Closer
 // interface if they haven't already been closed.
 func (c *Combiner) Close() error {
 	var err error
